Moving auth functions to authentication file
